% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpis.R
\name{get_dac}
\alias{get_dac}
\title{Get the average DAC value over a particular set of dates}
\usage{
get_dac(con, dates)
}
\arguments{
\item{con}{A connection to the \strong{intendo} database. We can use either of the
\code{\link[=db_con]{db_con()}} or \code{\link[=db_con_p]{db_con_p()}} functions to generate this.}

\item{dates}{A vector of dates.}
}
\description{
This function uses data from the \strong{intendo} database to get the DAC (daily
active customers) on a single day, or, the average over a set of days. You
need to supply the database connection object and a vector of dates in ISO
format.
}
